pyproject.toml
[project]
name = "bioresearch"
version = "0.1.0"
description = "Autonomous research engine for biology: perturbation prediction, ADMET profiling, clinical trial optimization"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
bio = [
    "scanpy>=1.10.0",
    "anndata>=0.10.0",
    "pertpy>=0.7.0",
]
chem = [
    "rdkit>=2023.9.1",
    "PyTDC>=1.0.0",
]
gpu = [
    "torch>=2.1.0",
]
web = [
    "gradio>=4.0.0",
]
infra = [
    "modal>=0.64.0",
]
dev = [
    "pytest>=7.0.0",
]
all = [
    "bioresearch[bio,chem,gpu,web,infra]",
]

[project.scripts]
bioresearch = "cli:main"

[tool.hatch.build.targets.wheel]
packages = ["engine", "infra", "knowledge", "domains", "web"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
